backend/requirements.txt
fastapi
uvicorn
numpy
soundfile
faster-whisper
resemblyzer
pydantic
pycryptodome
av
